mevbot/core/strategy.py
"""MEV strategy implementation"""
from typing import Dict, List, Optional, Tuple, Any
import numpy as np
from web3 import Web3
import logging
from eth_typing import Address
from dataclasses import dataclass
from concurrent.futures import ThreadPoolExecutor
import time

logger = logging.getLogger(__name__)

@dataclass
class OpportunityMetrics:
    profit: float
    gas_cost_usd: float
    success_probability: float
    execution_time_ms: float
    path: List[Address]
    required_loan: float = 0
    token_address: Optional[Address] = None

class MEVStrategy:
    def __init__(self, w3: Web3, config: Dict):
        self.w3 = w3
        self.config = config
        self.executor = ThreadPoolExecutor(max_workers=4)
        
        # SIMD optimization for calculations
        self.enable_numpy_optimization = True
        
    def calculate_profit(
        self,
        path: List[Address],
        amounts: np.ndarray,
        pool_states: Dict
    ) -> Tuple[np.ndarray, np.ndarray]:
        """Calculate profit for a given path and amounts"""
        if self.enable_numpy_optimization:
            # Calculate slippage for each pool in the path
            slippage = self._calculate_slippage_vectorized(amounts, pool_states, path)
            
            # Calculate gas costs (simplified)
            gas_costs = np.full_like(amounts, 0.001)  # 0.001 ETH per trade
            
            # Calculate profits
            profits = amounts * (1 - slippage) - gas_costs
            
            return profits, slippage
        else:
            # Fallback to regular calculation
            return self._calculate_profit_standard(path, amounts[0], pool_states)
    
    def _calculate_slippage_vectorized(
        self,
        amounts: np.ndarray,
        pool_states: Dict,
        path: List[Address]
    ) -> np.ndarray:
        """Vectorized slippage calculation using numpy"""
        if self.config.get('test_mode'):
            return np.full_like(amounts, 0.0005)  # 0.05% slippage in test mode
            
        # Get pool depths for each pool in the path
        pool_depths = np.array([
            pool_states[pool]['depth']
            for pool in path[:-1]  # Last address is output token
        ])
        
        # Reshape arrays for broadcasting
        amounts = amounts.reshape(-1, 1)  # Shape: (n_amounts, 1)
        pool_depths = pool_depths.reshape(1, -1)  # Shape: (1, n_pools)
        
        # Calculate slippage for each amount-pool combination
        slippage = np.sum(
            np.square(amounts) / pool_depths,
            axis=1  # Sum across pools
        )
        
        return slippage
        
    def calculate_success_probability(
        self,
        gas_price: int,
        network_congestion: float,
        competitor_count: int
    ) -> float:
        """Calculate probability of successful execution"""
        if self.config.get('test_mode'):
            return 0.9  # 90% success probability in test mode
            
        # Simple model: p = 1 / (1 + e^(-x))
        # where x depends on gas price, network congestion, and competition
        x = (
            0.1 * np.log(gas_price) -
            2.0 * network_congestion -
            0.5 * competitor_count
        )
        return float(1 / (1 + np.exp(-x)))
        
    def analyze_network_conditions(self) -> Dict:
        """Analyze current network conditions"""
        if self.config.get('test_mode'):
            return {
                'congestion': 0.5,
                'competitor_count': 5,
                'gas_price': 50000000000,  # 50 GWEI
                'block_number': 1000000  # Test block number
            }
            
        try:
            # Get latest block
            block = self.w3.eth.get_block('latest')
            
            # Calculate network congestion (0-1)
            target_gas = 15000000  # Target block gas
            congestion = min(1.0, block['gasUsed'] / target_gas)
            
            # Count active MEV bots (simplified)
            competitor_count = 10  # Placeholder
            
            # Get current gas price
            gas_price = block['baseFeePerGas']
            
            return {
                'congestion': congestion,
                'competitor_count': competitor_count,
                'gas_price': gas_price,
                'block_number': block['number']
            }
            
        except Exception as e:
            logger.error(f"Error analyzing network conditions: {e}")
            return {
                'congestion': 0.5,
                'competitor_count': 10,
                'gas_price': 50000000000,
                'block_number': 1000000  # Fallback block number
            }

class StrategyManager:
    def __init__(self, w3: Web3, config: Dict):
        self.w3 = w3
        self.config = config
        self.strategy = MEVStrategy(w3, config)
        self.active_opportunities: Dict[str, OpportunityMetrics] = {}
        
    async def analyze_opportunity(self, tx: Dict) -> Optional[OpportunityMetrics]:
        """Analyze a potential MEV opportunity from a transaction"""
        try:
            start_time = time.time()
            
            # Get network conditions
            conditions = self.strategy.analyze_network_conditions()
            
            # Create test path in test mode
            if self.config.get('test_mode'):
                path = [
                    Address('0x' + '1' * 40),
                    Address('0x' + '2' * 40),
                    Address('0x' + '3' * 40)
                ]
                pool_states = {
                    addr: {'depth': 1000000000000000000}  # 1 ETH depth
                    for addr in path
                }
            else:
                # Extract path from transaction
                path = self._extract_path(tx)
                if not path:
                    return None
                    
                # Get pool states
                pool_states = await self._get_pool_states(path)
            
            # Calculate profit using SIMD
            amounts = np.array([tx['value']])
            profits, slippage = self.strategy.calculate_profit(
                path, amounts, pool_states
            )
            
            # Calculate success probability
            success_prob = self.strategy.calculate_success_probability(
                conditions['gas_price'],
                conditions['congestion'],
                conditions['competitor_count']
            )
            
            # Create metrics
            metrics = OpportunityMetrics(
                profit=float(profits[0]),
                gas_cost_usd=float(conditions['gas_price'] * 21000 / 1e18),
                success_probability=success_prob,
                execution_time_ms=(time.time() - start_time) * 1000,
                path=path
            )
            
            # Store if profitable
            if self._is_profitable(metrics):
                opp_id = self._generate_opportunity_id(path)
                self.active_opportunities[opp_id] = metrics
            
            return metrics
            
        except Exception as e:
            logger.error(f"Error analyzing opportunity: {e}")
            return None
            
    def _is_profitable(self, metrics: OpportunityMetrics) -> bool:
        """Check if opportunity meets profit thresholds"""
        min_profit = self.config.get('min_profit', 0.001)  # 0.1% minimum profit
        min_success_prob = self.config.get('min_success_probability', 0.5)
        
        return (
            metrics.profit > min_profit and
            metrics.success_probability > min_success_prob
        )
        
    async def _get_pool_states(self, path: List[Address]) -> Dict:
        """Get current state of all pools in path"""
        if self.config.get('test_mode'):
            return {
                addr: {'depth': 1000000000000000000}  # 1 ETH depth
                for addr in path
            }
            
        # Implement actual pool state fetching here
        return {}
        
    def _generate_opportunity_id(self, path: List[Address]) -> str:
        """Generate unique ID for opportunity"""
        return '_'.join(str(addr) for addr in path)
        
    def _extract_path(self, tx: Dict) -> Optional[List[Address]]:
        """Extract trading path from transaction"""
        if self.config.get('test_mode'):
            return [
                Address('0x' + '1' * 40),
                Address('0x' + '2' * 40),
                Address('0x' + '3' * 40)
            ]
            
        # Implement actual path extraction here
        return None
